% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_indel.R
\name{AssignIndelPCAWGClassification}
\alias{AssignIndelPCAWGClassification}
\title{Assigns PCAWG classification of an indel}
\usage{
AssignIndelPCAWGClassification(x)
}
\arguments{
\item{x}{One row of data.frame the function PreprocessIndel.}
}
\value{
A data.frame with the following columns:
\item{PCAWG_Indel_Class}{PCAWG indel mutation type (e.g. 'DEL_C_1_1').}
\item{Microhomology_Class}{PCAWG microhomology mutation type (e.g. 'DEL_MH_2_1').}
\item{Microhomology_Size}{Microhomology size.}
\item{Microhomology_Seq}{Microhomology sequence.}
\item{Microhomology_Direction}{Microhomology direction.}
}
\description{
This function returns PCAWG classification data of an indel.
}
