% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maximum_likelihood_estimation.R
\name{EstimateMaximumLikelihood}
\alias{EstimateMaximumLikelihood}
\title{Performs maximum likelihood estimation}
\usage{
EstimateMaximumLikelihood(
  mutation.frequencies,
  signature.weights,
  num.mutations,
  min.probability = 0.01
)
}
\arguments{
\item{mutation.frequencies}{A numeric vector corresponding to the frequency of each mutation type}

\item{signature.weights}{A matrix of signature weights.}

\item{num.mutations}{An integer value that specifies the number of mutations}

\item{min.probability}{A float value that specifies the minimum probability to attribute to a signature.}
}
\value{
A list with the following items:
\item{Cosine_Similarity}{Cosine similarity score between mutation.frequncies and Reconstructed_Spectrum}
\item{Par}{A numeric vector of contribution weights of identified signatures.}
\item{Par_Normalized}{A normalized (0 to 1) numeric vector of contribution weights of identified signatures.}
\item{Reconstructed_Spectrum}{A numeric vector of reconstructed spectrum.}
\item{Reconstructed_Spectrum_Normalized}{A normalized (0 to 1) numeric vector of reconstructed spectrum.}
\item{Residuals}{A numeric vector of residual spectrum.}
\item{Residuals_Normalized}{A normalized (0 to 1) numeric vector of residual spectrum.}
\item{RSS}{Residual sum of squares (of Residuals).}
\item{RSS_Normalized}{Residual sum of squares (of Residuals_Normalized).}
}
\description{
This function performs maximum likelihood estimation of a given identification setup.
}
