% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{GetComplementBase}
\alias{GetComplementBase}
\title{Returns complement base}
\usage{
GetComplementBase(nucleotide)
}
\arguments{
\item{nucleotide}{Nucleotide base.}
}
\value{
"A", "C", "G", or "T"
}
\description{
This function returns the complement base of a nucleotide
}
