% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{GetComplementBases}
\alias{GetComplementBases}
\title{Returns complement bases}
\usage{
GetComplementBases(nucleotides)
}
\arguments{
\item{nucleotides}{Nucleotide bases (e.g. "AG").}
}
\value{
e.g. "TC"
}
\description{
This function returns the complement bases of a nucleotide sequence
}
