% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/housekeeping.R
\name{GetCosmicDbsSignaturesData}
\alias{GetCosmicDbsSignaturesData}
\title{Returns the COSMIC v3.1 DBS reference signature data}
\usage{
GetCosmicDbsSignaturesData()
}
\value{
A data.frame of the PCAWG DBS mutational signatures.
}
\description{
This function returns the COSMIC v3.1 doublet base substitution
(DBS) mutational signatures data.
}
