% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/housekeeping.R
\name{GetCosmicDbsSignaturesNames}
\alias{GetCosmicDbsSignaturesNames}
\title{Return the COSMIC v3.1 DBS reference signature names}
\usage{
GetCosmicDbsSignaturesNames()
}
\value{
Returns a character vector of all COSMIC v3.1 DBS mutational signature names.
}
\description{
This function returns the COSMIC v3.1 DBS mutational signatures names.
}
