% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/housekeeping.R
\name{GetCosmicIndelSignaturesData}
\alias{GetCosmicIndelSignaturesData}
\title{Returns the COSMIC v3.1 indel reference signature data}
\usage{
GetCosmicIndelSignaturesData()
}
\value{
A data.frame of the COSMIC v3.1 indel mutational signatures.
}
\description{
This function returns the COSMIC v3.1 small insertion and deletion
(indel) mutational signatures data.
}
