% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/housekeeping.R
\name{GetCosmicIndelSignaturesNames}
\alias{GetCosmicIndelSignaturesNames}
\title{Returns the COSMIC v3.1 indel signature names}
\usage{
GetCosmicIndelSignaturesNames()
}
\value{
Returns a character vector of all COSMIC v3.1 indel mutational signature names.
}
\description{
This function returns the COSMIC v3.1 indel mutational signatures names.
}
