% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/housekeeping.R
\name{GetCosmicSbsSignaturesData}
\alias{GetCosmicSbsSignaturesData}
\title{Returns the COSMIC v3.1 SBS reference signature data}
\usage{
GetCosmicSbsSignaturesData()
}
\value{
A data.frame of the COSMIC v3.1 SBS mutational signatures.
}
\description{
This function returns the COSMIC v3.1 single base substitution
(SBS) mutational signatures data.
}
