% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/housekeeping.R
\name{GetCosmicSbsSignaturesNames}
\alias{GetCosmicSbsSignaturesNames}
\title{Return the COSMIC v3.1 SBS reference signature names}
\usage{
GetCosmicSbsSignaturesNames()
}
\value{
Returns a character vector of all COSMIC v3.1 SBS mutational signature names.
}
\description{
This function returns the COSMIC v3.1 SBS mutational signatures names.
}
