% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/housekeeping.R
\name{GetDbsSignaturesColors}
\alias{GetDbsSignaturesColors}
\title{Returns the DBS signatures colors and etiologies}
\usage{
GetDbsSignaturesColors()
}
\value{
A data.frame of the DBS mutational signatures colors and etiologies.
}
\description{
This function returns the DBS mutational signatures colors and etiologies.
}
