% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_theme.R
\name{GetDbsSignaturesPlotTheme}
\alias{GetDbsSignaturesPlotTheme}
\title{Fetches DBS plotting theme}
\usage{
GetDbsSignaturesPlotTheme(
  x.axis.labels = PLOT.DBS.X.AXIS.LABELS,
  x.axis.text.size = 10,
  x.axis.text.angle = 90,
  x.axis.text.hjust = 1,
  x.axis.text.vjust = 0.5,
  y.axis.text.size = 12,
  y.axis.title.size = 12,
  y.axis.min = 0,
  y.axis.max = 1,
  y.axis.title = "Mutation Fraction",
  groups = DBS.MUTATION.TYPES.GROUPS,
  strip.labels = PLOT.DBS.STRIP.LABELS,
  strip.colors = PLOT.DBS.STRIP.COLORS,
  strip.text.colors = PLOT.DBS.STRIP.TEXT.COLORS,
  strip.text.size = 12,
  legend.title = "Mutation Type",
  legend.title.size = 12,
  legend.text.size = 10,
  legend.labels = PLOT.DBS.LEGEND.LABELS
)
}
\arguments{
\item{x.axis.labels}{A character vector used to label the x-axis.}

\item{x.axis.text.size}{x-axis text size.}

\item{x.axis.text.angle}{x-axis text angle.}

\item{y.axis.text.size}{y-axis text size.}

\item{y.axis.title.size}{y-axis title size.}

\item{y.axis.min}{y-axis lower-bound (minimum range for limits).}

\item{y.axis.max}{y-axis upper-bound (maximum range for limits).}

\item{y.axis.title}{y-axis title text.}

\item{groups}{A character vector of group variable values.}

\item{strip.labels}{A character vector used to label the strip.}

\item{strip.colors}{A character vector of hex values used to color the strip.}

\item{strip.text.colors}{A character vector of hex values used to color the strip text.}

\item{strip.text.size}{Strip (facet) text size.}

\item{legend.title}{Legend title text.}

\item{legend.title.size}{Legend title text size.}

\item{legend.labels}{A character vector used to label the legend items.}
}
\value{
A data.frame of the DBS signature plot theme elements.
}
\description{
This function returns the DBS plotting theme elements.
}
