% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_indel.R
\name{GetDeletionMicrohomologySize}
\alias{GetDeletionMicrohomologySize}
\title{Returns microhomology size of a deletion}
\usage{
GetDeletionMicrohomologySize(upstream.seq, downstream.seq, mutated.seq)
}
\arguments{
\item{upstream.seq}{Upstream nucleotide sequence.}

\item{downstream.seq}{Downstream nucleotide sequence.}

\item{mutated.seq}{Indel nucleotide sequence}
}
\value{
A data.frame with the following columns:
\item{Microhomology_Seq}{Microhomology nucleotide sequence.}
\item{Microhomology_Size}{Size (number of bases) of microhomology.}
\item{Microhomology_Direction}{Direction of microhomology.}
}
\description{
This function computes and returns the microhomology size of a small deletion.
}
