% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/housekeeping.R
\name{GetHsapiensEnsemblData}
\alias{GetHsapiensEnsemblData}
\title{Returns the hsapiens ensembl data}
\usage{
GetHsapiensEnsemblData(version)
}
\arguments{
\item{version}{Either "GRCh37" or "GRCh38".}
}
\description{
This function returns the hsapiens ensembl data.
}
