% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_indel.R
\name{GetIndelPCAWGClassification}
\alias{GetIndelPCAWGClassification}
\title{Returns PCAWG classification of an indel}
\usage{
GetIndelPCAWGClassification(type, mutated.seq, upstream.seq, downstream.seq)
}
\arguments{
\item{type}{Indel type. Either 'INS' or 'DEL'.}

\item{mutated.seq}{Indel nucleotide sequence.}

\item{upstream.seq}{Upstream nucleotide sequence.}

\item{downstream.seq}{Downstream nucleotide sequence.}
}
\value{
A data.frame with the following columns:
\item{PCAWG_Indel_Class}{PCAWG indel mutation type (e.g. 'DEL_C_1_0').}
\item{Microhomology_Class}{PCAWG microhomology mutation type (e.g. 'DEL_MH_2_1').}
\item{Microhomology_Size}{Microhomology size.}
\item{Microhomology_Seq}{Microhomology sequence.}
\item{Microhomology_Direction}{Microhomology direction.}
}
\description{
This function classifies an indel according to the PCAWG classification.
}
