% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_indel.R
\name{GetIndelRepeatSize}
\alias{GetIndelRepeatSize}
\title{Returns indel repeat size}
\usage{
GetIndelRepeatSize(downstream.seq, mutated.seq, type)
}
\arguments{
\item{downstream.seq}{Downstream nucleotide sequence.}

\item{mutated.seq}{Indel nucleotide sequence.}

\item{type}{Indel type. Either 'INS' or 'DEL'.}
}
\value{
Repeat size (integer).
}
\description{
This function computes and returns the repeat size of an indel variant.
}
