% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/housekeeping.R
\name{GetIndelSignaturesColors}
\alias{GetIndelSignaturesColors}
\title{Returns the indel signatures colors and etiologies}
\usage{
GetIndelSignaturesColors()
}
\value{
A data.frame of the indel mutational signatures colors and etiologies.
}
\description{
This function returns the indel mutational signatures colors and etiologies.
}
