% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/housekeeping.R
\name{GetKucabSbsSignaturesData}
\alias{GetKucabSbsSignaturesData}
\title{Returns the Kucab SBS reference signature data}
\usage{
GetKucabSbsSignaturesData()
}
\value{
A data.frame of the Kucab SBS mutational signatures.
}
\description{
This function returns the Kucab single base subsitution
(SBS) mutational signatures data.
}
