% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/housekeeping.R
\name{GetKucabSbsSignaturesNames}
\alias{GetKucabSbsSignaturesNames}
\title{Return the Kucab SBS reference signature names}
\usage{
GetKucabSbsSignaturesNames()
}
\description{
This function returns the Kucab SBS mutational signatures names.
}
