% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/housekeeping.R
\name{GetMmusculusEnsemblData}
\alias{GetMmusculusEnsemblData}
\title{Returns the mmusculus ensembl data}
\usage{
GetMmusculusEnsemblData(version)
}
\arguments{
\item{version}{Either "NCBIM37" or "GRCm38".}
}
\description{
This function returns the mmusculus ensembl data.
}
