% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/housekeeping.R
\name{GetPcawgDbsSignaturesData}
\alias{GetPcawgDbsSignaturesData}
\title{Returns the PCAWG DBS reference signature data}
\usage{
GetPcawgDbsSignaturesData()
}
\value{
A data.frame of the PCAWG DBS mutational signatures.
}
\description{
This function returns the PCAWG doublet base subsitution
(DBS) mutational signatures data.
}
