% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/housekeeping.R
\name{GetPcawgDbsSignaturesNames}
\alias{GetPcawgDbsSignaturesNames}
\title{Return the DBS reference signature names}
\usage{
GetPcawgDbsSignaturesNames()
}
\value{
Returns a character vector of all PCAWG DBS mutational signature names.
}
\description{
This function returns the PCAWG DBS mutational signatures names.
}
