% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/housekeeping.R
\name{GetPcawgIndelSignaturesData}
\alias{GetPcawgIndelSignaturesData}
\title{Returns the PCAWG indel reference signature data}
\usage{
GetPcawgIndelSignaturesData(version)
}
\arguments{
\item{version}{Either "SigProfiler" or "SignatureAnalyzer".}
}
\value{
A data.frame of the PCAWG indel mutational signatures.
}
\description{
This function returns the PCAWG small insertion and deletion
(indel) mutational signatures data.
}
