% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/housekeeping.R
\name{GetPcawgIndelSignaturesNames}
\alias{GetPcawgIndelSignaturesNames}
\title{Returns the PCAWG indel signature names}
\usage{
GetPcawgIndelSignaturesNames(version)
}
\arguments{
\item{version}{Either "SigProfiler" or "SignatureAnalyzer".}
}
\value{
Returns a character vector of all PCAWG indel mutational signature names.
}
\description{
This function returns the PCAWG indel mutational signatures names.
}
