% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/housekeeping.R
\name{GetPcawgPlatinumSbsSignaturesData}
\alias{GetPcawgPlatinumSbsSignaturesData}
\title{Returns the PCAWG Platinum SBS reference signature data}
\usage{
GetPcawgPlatinumSbsSignaturesData()
}
\value{
A data.frame of the PCAWG Platinum SBS mutational signatures.
}
\description{
This function returns the PCAWG Platinum single base subsitution
(SBS) mutational signatures data (from Petljak et al., Cell 2019).
}
