% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/housekeeping.R
\name{GetPcawgPlatinumSbsSignaturesNames}
\alias{GetPcawgPlatinumSbsSignaturesNames}
\title{Return the PCAWG Platinum SBS reference signature names}
\usage{
GetPcawgPlatinumSbsSignaturesNames()
}
\value{
Returns a character vector of all PCAWG Platinum SBS mutational signature names.
}
\description{
This function returns the PCAWG Platinum SBS mutational signatures names.
}
