% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/housekeeping.R
\name{GetPcawgSbsPentanucleotideSignaturesData}
\alias{GetPcawgSbsPentanucleotideSignaturesData}
\title{Returns the PCAWG SBS penta-nucleotide reference signature data}
\usage{
GetPcawgSbsPentanucleotideSignaturesData()
}
\value{
A data.frame of the PCAWG SBS penta-nucleotide mutational signatures.
}
\description{
This function returns the PCAWG single base substitution
(SBS) pentanucleotide context mutational signatures data.
}
