% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/housekeeping.R
\name{GetPcawgSbsPentanucleotideSignaturesNames}
\alias{GetPcawgSbsPentanucleotideSignaturesNames}
\title{Return the PCAWG SBS penta-nucleotide reference signature names}
\usage{
GetPcawgSbsPentanucleotideSignaturesNames()
}
\value{
Returns a character vector of all PCAWG SBS penta-nucleotide mutational signature names.
}
\description{
This function returns the PCAWG SBS penta-nucleotide mutational signatures names.
}
