% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/housekeeping.R
\name{GetPcawgSbsSignaturesData}
\alias{GetPcawgSbsSignaturesData}
\title{Returns the PCAWG SBS reference signature data}
\usage{
GetPcawgSbsSignaturesData(sequencing.type)
}
\arguments{
\item{sequencing.type}{Either 'WES' for whole-exome sequencing or
'WGS' for whole-genome sequencing.}
}
\value{
A data.frame of the PCAWG SBS mutational signatures.
}
\description{
This function returns the PCAWG single base subsitution
(SBS) mutational signatures data.
}
