% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/housekeeping.R
\name{GetPcawgSbsSignaturesNames}
\alias{GetPcawgSbsSignaturesNames}
\title{Return the PCAWG SBS reference signature names}
\usage{
GetPcawgSbsSignaturesNames(sequencing.type)
}
\arguments{
\item{sequencing.type}{Either 'WES' for whole-exome sequencing or
'WGS' for whole-genome sequencing.}
}
\value{
Returns a character vector of all PCAWG SBS mutational signature names.
}
\description{
This function returns the PCAWG SBS mutational signatures names.
}
