% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/housekeeping.R
\name{GetPetljakSbsSignaturesData}
\alias{GetPetljakSbsSignaturesData}
\title{Returns the Petljak SBS reference signature data}
\usage{
GetPetljakSbsSignaturesData()
}
\value{
A data.frame of the Petljak SBS mutational signatures.
}
\description{
This function returns the Petljak single base subsitution
(SBS) mutational signatures data.
}
