% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/housekeeping.R
\name{GetPetljakSbsSignaturesNames}
\alias{GetPetljakSbsSignaturesNames}
\title{Return the Petljak SBS reference signature names}
\usage{
GetPetljakSbsSignaturesNames(sequencing.type)
}
\value{
Returns a character vector of all Petljak SBS mutational signature names.
}
\description{
This function returns the Petljak SBS mutational signatures names.
}
