% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/housekeeping.R
\name{GetPleguezuelosManzanoIndelSignaturesData}
\alias{GetPleguezuelosManzanoIndelSignaturesData}
\title{Returns the Pleguezuelos-Manzano indel reference signature data}
\usage{
GetPleguezuelosManzanoIndelSignaturesData()
}
\value{
A data.frame of the Pleguezuelos-Manzano indel mutational signatures.
}
\description{
This function returns the Pleguezuelos-Manzano small insertion and deletion
(indel) mutational signatures data.
}
