% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/housekeeping.R
\name{GetPleguezuelosManzanoIndelSignaturesNames}
\alias{GetPleguezuelosManzanoIndelSignaturesNames}
\title{Returns the Pleguezuelos-Manzano indel signature names}
\usage{
GetPleguezuelosManzanoIndelSignaturesNames()
}
\value{
Returns a character vector of all Pleguezuelos-Manzano indel mutational signature names.
}
\description{
This function returns the Pleguezuelos-Manzano indel mutational signatures names.
}
