% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/housekeeping.R
\name{GetPleguezuelosManzanoSbsSignaturesData}
\alias{GetPleguezuelosManzanoSbsSignaturesData}
\title{Returns the Pleguezuelos-Manzano SBS reference signature data}
\usage{
GetPleguezuelosManzanoSbsSignaturesData()
}
\value{
A data.frame of the Petljak Pleguezuelos-Manzano mutational signatures.
}
\description{
This function returns the Pleguezuelos-Manzano single base subsitution
(SBS) mutational signatures data.
}
