% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/housekeeping.R
\name{GetPleguezuelosManzanoSbsSignaturesNames}
\alias{GetPleguezuelosManzanoSbsSignaturesNames}
\title{Return the Pleguezuelos-Manzano SBS reference signature names}
\usage{
GetPleguezuelosManzanoSbsSignaturesNames()
}
\description{
This function returns the Pleguezuelos-Manzano
SBS mutational signatures names.
}
