% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/housekeeping.R
\name{GetSbsSignaturesColors}
\alias{GetSbsSignaturesColors}
\title{Returns the SBS signatures colors and etiologies}
\usage{
GetSbsSignaturesColors()
}
\value{
A data.frame of the SBS mutational signatures colors and etiologies.
}
\description{
This function returns the SBS mutational signatures colors and etiologies.
}
