% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{IdentifyDbsSignatures}
\alias{IdentifyDbsSignatures}
\title{Identify doublet base subsitution mutational signatures}
\usage{
IdentifyDbsSignatures(
  input,
  bsg,
  sample.id = "Sample",
  reference = GetPcawgDbsSignaturesData(),
  target.signatures = GetPcawgDbsSignaturesNames(),
  plot.theme = GetDbsSignaturesPlotTheme(),
  analyze.variants.column.gene,
  analyze.variants.column.group,
  analyze.variants = FALSE,
  n.cores = 2,
  combn.m = 3,
  n.max.signatures = 7,
  min.probability = 0.01,
  zeta.value = 1e-10,
  save = TRUE,
  save.dir = NULL
)
}
\arguments{
\item{input}{Either VCF file path or data.frame. If the input is a data.frame, it must include the following columns: Chr, Pos, Ref, Alt.}

\item{bsg}{BSgenome object.}

\item{sample.id}{Sample ID that will be used to name output files (default: 'Sample').}

\item{reference}{A data.frame with the following columns: Mutation_Type, and names of DBS signatures
(default: a data.frame returned from \code{\link{GetPcawgDbsSignaturesData}}).}

\item{target.signatures}{Signatures to be considered for identification (default: an array returned from \code{\link{GetPcawgDbsSignaturesNames}}).}

\item{plot.theme}{A data.frame returned from \code{\link{GetDbsSignaturesPlotTheme}}.}

\item{analyze.variants.column.gene}{Name of column in the data.frame corresponding to the gene name or ID (e.g. "Gene.refGene" if using ANNOVAR data).}

\item{analyze.variants.column.group}{Name of column in the data.frame corresponding to the variant group for plotting purposes (e.g. "Func.refGene" if using ANNOVAR data).}

\item{n.cores}{Number of cores to use.}

\item{combn.m}{Number of signatures to consider in each step. 'm' parameter in combn function (default: 3).}

\item{n.max.signatures}{Maximum number of signatures to identify. Recommended: n.max.signatures >= initial.exploration.combn.m (default: 7).}

\item{min.probability}{Minimum probability to attribute to a signature (default: 0.01).}

\item{zeta.value}{A float value that is added to the data frequency (default: 1e-10).}

\item{save}{Save resulting files if TRUE, otherwise do not save (default: TRUE).}

\item{save.dir}{Save directory path (default: NULL).}
}
\value{
A list with the following elements:

results: a data.frame with the following columns:
\item{Mutations_Count}{Number of mutations.}
\item{Signatures}{Identified mutational signatures separated by comma.}
\item{Signatures_Count}{Number of identified mutational signatures.}
\item{Signatures_Weights}{Normalized (0 to 1) weights of identified mutational signatures separated by comma.}
\item{Mutation_Types}{Mutation types separated by comma.}
\item{Mutation_Types_Groups}{Mutation type groups separated by comma.}
\item{Observed_Spectrum}{Normalized spectrum (frequency) of observed mutations separated by comma.}
\item{Reconstructed_Spectrum}{Normalized spectrum (frequency) of MLE reconstructed mutations separated by comma.}
\item{Residual_Spectrum}{Normalized spectrum (frequency) of residual mutations separated by comma.}
\item{Cosine_Similarity}{Cosine similarity score between Observed_Spectrum and Reconstructed_Spectrum.}
\item{RSS}{Raw residual sum of squares (derived from Residual_Spectrum).}
\item{RSS_Normalized}{Normalized residual sum of squares (derived from Residual_Spectrum).}
\item{BIC}{Bayesian information criterion of the identified model.}

results.variants: a data.frame with the following columns:
}
\description{
This function identifies DBS mutational signatures.
}
