% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signature_identification.R
\name{IdentifySignatures}
\alias{IdentifySignatures}
\title{Identify contribution weights of underlying mutational signatures}
\usage{
IdentifySignatures(
  data,
  reference,
  target.signatures,
  n.cores = 2,
  combn.m = 3,
  n.max.signatures = 7,
  min.probability = 0.01,
  zeta.value = 1e-10
)
}
\arguments{
\item{data}{A data.frame prepared by either PrepareSbsVcfFile, PrepareDbsVcfFile, PrepareIdVcfFile, or PrepareSvVcfFile depending on signature.type.}

\item{reference}{A data.frame where the first column is the list of mutation types and where 2nd to k-th column headers are the names of the signatures.}

\item{target.signatures}{A character vector that lists all desired signatures to consider for identification.}

\item{n.cores}{Number of cores to use (default: 2).}

\item{combn.m}{An integer value that specifies the number of signatures to consider in each step. 'm' parameter in combn function.}

\item{n.max.signatures}{An integer value that specifies the maximum number of signatures to consider. It is recommended that n.max.signatures >= initial.exploration.combn.m.}

\item{min.probability}{A float value that specifies the minimum probability to attribute to a signature.}

\item{zeta.value}{A float value that is added to the data frequency.}
}
\value{
A data.frame with the following columns:
\item{Mutations_Count}{Number of mutations.}
\item{Signatures}{Identified mutational signatures separated by comma.}
\item{Signatures_Count}{Number of identified mutational signatures.}
\item{Signatures_Weights}{Normalized (0 to 1) weights of identified mutational signatures separated by comma.}
\item{Mutation_Types}{Mutation types separated by comma.}
\item{Mutation_Types_Groups}{Mutation type groups separated by comma.}
\item{Observed_Spectrum}{Normalized spectrum (frequency) of observed mutations separated by comma.}
\item{Reconstructed_Spectrum}{Normalized spectrum (frequency) of MLE reconstructed mutations separated by comma.}
\item{Residual_Spectrum}{Normalized spectrum (frequency) of residual mutations separated by comma.}
\item{Cosine_Similarity}{Cosine similarity score between Observed_Spectrum and Reconstructed_Spectrum.}
\item{RSS}{Raw residual sum of squares (derived from Residual_Spectrum).}
\item{RSS_Normalized}{Normalized residual sum of squares (derived from Residual_Spectrum).}
\item{BIC}{Bayesian information criterion of the identified model.}
}
\description{
This function identifies which mutational signatures are underlying a
given data.frame and contribution weights of each mutational signature.
}
