% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variant_signature.R
\name{IdentifyVariantSignatures}
\alias{IdentifyVariantSignatures}
\title{Identifies mutational signatures associated with each variant}
\usage{
IdentifyVariantSignatures(prepared.data, identified.model, reference)
}
\arguments{
\item{prepared.data}{A data.frame returned from \code{\link{PrepareSbsDataFrame}}.}

\item{identified.model}{A data.frame returned from \code{\link{IdentifySignatures}}.}

\item{reference}{A data.frame with the following columns: Mutation_Type, and names of SBS signatures.}
}
\value{
A data.frame with the following columns:
\item{Chr}{Chromosome.}
\item{Pos}{Genomic position.}
\item{Ref}{Reference allele.}
\item{Alt}{Alternate allele.}
\item{Mutation_Type}{Mutation type.}
\item{Variant_Gene}{Variant gene.}
\item{Variant_Group}{Variant group.}
\item{Signature}{Signature.}
\item{Probability}{Probability the mutation is the result of the signature.}
}
\description{
This function identifies mutational signatures associated with each variant.
}
