% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maximum_likelihood_estimation.R
\name{LinearRegressionOptimFn}
\alias{LinearRegressionOptimFn}
\title{Computes and returns the residual sum of squares}
\usage{
LinearRegressionOptimFn(par, mutation.frequencies, signature.weights)
}
\arguments{
\item{par}{A numeric vector of parameters to optimze.}

\item{mutation.frequencies}{A numeric vector corresponding to the frequency of each mutation type.}

\item{signature.weights}{A matrix of signature weights.}
}
\value{
Residual sum of squares.
}
\description{
This function estimates the residual sum of squares.
}
