% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{PlotIdentifiedModel}
\alias{PlotIdentifiedModel}
\title{Plots identified model}
\usage{
PlotIdentifiedModel(identified.model, reference, plot.theme)
}
\arguments{
\item{identified.model}{A row of the data.frame returned from the function IdentifySignatures.}

\item{reference}{A data.frame with the following columns: Mutation_Type, and names of signatures.}

\item{plot.theme}{A data.frame returned from \code{\link{GetSbsSignaturesPlotTheme}},
\code{\link{GetDbsSignaturesPlotTheme}} or \code{\link{GetIndelSignaturesPlotTheme}}.}
}
\value{
A list with the following items:
\item{plot.merged}{Merged (plot.observed, plot.reconstructed.spectrum, plot.residual.spectrum, plot.legend) plot.}
\item{plot.observed}{Observed mutation frequencies plot.}
\item{plot.reconstructed.spectrum}{Reconstructed mutation frequencies plot.}
\item{plot.residual.spectrum}{Residual mutation frequencies plot.}
\item{plot.legend}{Legend plot.}
}
\description{
This function plots an identified model.
}
