% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_group.R
\name{PlotIdentifiedModels}
\alias{PlotIdentifiedModels}
\title{Plots identified models}
\usage{
PlotIdentifiedModels(
  df.models,
  df.signatures.colors,
  signature.group.order = c(),
  sample.ids.order = c(),
  apply.facet = FALSE,
  df.sample.ids.groups = c(),
  sample.ids.groups.order = c(),
  multiply.by.mutations.count = FALSE,
  y.axis.max = NULL,
  plot.title = "",
  title.size = 12,
  x.axis.text.size = 8,
  y.axis.text.size = 10,
  y.axis.title.size = 12,
  legend.title.size = 12,
  legend.text.size = 10
)
}
\arguments{
\item{df.models}{A data.frame of all best models appended by rbind with Sample_ID added as a column.}

\item{df.signatures.colors}{A data.frame returned from
\code{\link{GetSbsSignaturesColors}}, \code{\link{GetDbsSignaturesColors}}, or \code{\link{GetIndelSignaturesColors}}.}

\item{signature.group.order}{A character vector specifying the order of the signature groups.}

\item{x.axis.text.size}{A numeric value that sets the x.axis.text size.}

\item{y.axis.text.size}{A numeric value that sets the y.axis.text size.}

\item{y.axis.title.size}{A numeric value that sets the y.axis.title size.}

\item{legend.title.size}{A numeric value that sets the legend title size.}

\item{legend.text.size}{A numeric value that sets the legend text size.}
}
\value{
A ggplot object.
}
\description{
This function plots identified models for a group of samples IDs.
}
