% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{PlotIdentifiedModelSignatureWeights}
\alias{PlotIdentifiedModelSignatureWeights}
\title{Plots signature weights of an identified model}
\usage{
PlotIdentifiedModelSignatureWeights(
  identified.model,
  df.signatures.colors,
  plot.theme
)
}
\arguments{
\item{identified.model}{A row of the data.frame returned from the function IdentifySignatures.}

\item{df.signature.colors}{A data.frame with the following columns: Mutation_Type, and names of signatures.}
}
\value{
ggplot object.
}
\description{
This function plots signature weights of an identified model.
}
