% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_group.R
\name{PlotIdentifiedModelVariantSignatures}
\alias{PlotIdentifiedModelVariantSignatures}
\title{Plots a stacked barplot of signatures probabilities of variants}
\usage{
PlotIdentifiedModelVariantSignatures(df.sigs.probs, df.signatures.colors)
}
\value{
A list with the following items:
\item{plots.merged}{Merged rainfall plots.}
\item{plots}{Rainfall plot for each chromosome.}
}
\description{
This function plots a stacked barplot of signature probabilities of variants.
}
