% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{PlotSpectrum}
\alias{PlotSpectrum}
\title{Plots spectrum}
\usage{
PlotSpectrum(y, plot.theme)
}
\arguments{
\item{y}{A numeric vector corresponding to the y-axis values.}

\item{plot.theme}{Plot theme (returned from GetSbsSignaturesPlotTheme).}
}
\value{
A list with the following items:
\item{plot}{plot.}
\item{legend}{legend plot.}
}
\description{
This function plots spectrum of mutation frequencies.
}
