% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annovar_parsing.R
\name{PrepareAnnovarFile}
\alias{PrepareAnnovarFile}
\title{Fetches the closest Gene.refGene from an ANNOVAR file}
\usage{
PrepareAnnovarFile(annovar.file)
}
\arguments{
\item{annovar.file}{ANNOVAR file including path.}
}
\value{
A data.frame with the following items:
\item{Chr}{Chromosome name.}
\item{Pos}{Genomic position.}
\item{Ref}{Nucleotide(s).}
\item{Alt}{Nucleotide(s).}
\item{Gene.refGene}{Closest Gene.refGene.}
\item{Func.refGene}{Functional consequence of variant.}
}
\description{
This function prepares an ANNOVAR file by fetching the closest Gene.refGene from an ANNOVAR file.
If the genomic distance information cannot be parsed, the original 'Gene.refGene' value is returned.
}
