% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_dbs.R
\name{PrepareDbsDataFrame}
\alias{PrepareDbsDataFrame}
\title{Prepares DBS data.frame}
\usage{
PrepareDbsDataFrame(df, bsg, analyze.variants)
}
\arguments{
\item{df}{A data.frame with the following columns: Chr, Pos, Ref, Alt.}

\item{bsg}{BSgenome object.}

\item{analyze.variants}{A boolean value that indicates whether variant-level signature analysis should be performed.}
}
\value{
A list with the following items:
\item{df.dbs}{A data.frame of DBS data.}
\item{df.dbs.frequencies}{A data.frame of DBS mutation tyep frequencies.}
}
\description{
This function prepares a DBS data.frame and
returns relevant data to run IdentifySignatures function.
}
