% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_dbs.R
\name{PrepareDbsVcfFile}
\alias{PrepareDbsVcfFile}
\title{Prepares DBS VCF file}
\usage{
PrepareDbsVcfFile(vcf.file, bsg)
}
\arguments{
\item{vcf.file}{VCF file including path.}

\item{bsg}{BSgenome object.}
}
\value{
A list with the following items:
\item{df.dbs}{A data.frame of DBS data.}
\item{df.dbs.frequencies}{A data.frame of DBS mutation type frequencies.}
}
\description{
This function prepares a DBS VCF file and
returns relevant data to run IdentifySignatures function.
}
