% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_indel.R
\name{PrepareIndelDataFrame}
\alias{PrepareIndelDataFrame}
\title{Prepares indel data.frame}
\usage{
PrepareIndelDataFrame(
  df,
  bsg,
  analyze.variants,
  max.len = 25,
  padding.len = 80
)
}
\arguments{
\item{df}{A data.frame with the following columns: Chr, Pos, Ref, Alt.}

\item{bsg}{BSgenome object.}

\item{analyze.variants}{A boolean value that indicates whether variant-level signature analysis should be performed.}

\item{max.len}{Maximum length of an indel to include. Indels longer than this length will be excluded.}

\item{padding.len}{Number of bases to pad upstream and downstream of each indel.}
}
\value{
A list with the following items:
\item{df.indel}{A data.frame of the indel data.}
\item{df.indel.frequencies}{A data.frame of the PCAWG indel class frequencies.}
}
\description{
This function prepares an indel data.frame and
returns relevant data to run IdentifySignatures function.
}
