% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_indel.R
\name{PrepareIndelVcfFile}
\alias{PrepareIndelVcfFile}
\title{Prepares indel VCF file}
\usage{
PrepareIndelVcfFile(vcf.file, bsg, max.len = 25, padding.len = 80)
}
\arguments{
\item{vcf.file}{VCF file including path.}

\item{bsg}{BSgenome object.}

\item{max.len}{Maximum length of an indel to include. Indels longer than this length will be excluded.}

\item{padding.len}{Number of bases to pad upstream and downstream of each indel.}
}
\value{
A list with the following items:
\item{df.indel}{A data.frame of the indel data.}
\item{df.indel.frequencies}{A data.frame of the PCAWG indel class frequencies.}
}
\description{
This function prepares an indel VCF file and
returns relevant data to run IdentifySignatures function.
}
