% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_sbs.R
\name{PrepareSbsDataFrame}
\alias{PrepareSbsDataFrame}
\title{Prepares SBS data.frame}
\usage{
PrepareSbsDataFrame(df, bsg, reference, context.length, analyze.variants)
}
\arguments{
\item{df}{A data.frame with the following columns: Chr, Pos, Ref, Alt.}

\item{bsg}{BSgenome object.}

\item{reference}{A data.frame with the following columns: Mutation_Type, and names of SBS signatures (default: a data.frame returned from \code{\link{GetPcawgSbsSignaturesData}}).}

\item{context.length}{Number of context nucleotides.}

\item{analyze.variants}{A boolean value that indicates whether variant-level signature analysis should be performed (default: FALSE).}
}
\value{
A list with the following items:
\item{df.sbs}{A data.frame of SBS data.}
\item{df.sbs.frequencies}{A data.frame of SBS mutation tyep frequencies.}
}
\description{
This function prepares a SBS data.frame and
returns relevant data to run IdentifySignatures function.
}
