% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_sbs.R
\name{PrepareSbsVcfFile}
\alias{PrepareSbsVcfFile}
\title{Prepares SBS VCF file}
\usage{
PrepareSbsVcfFile(vcf.file, bsg, reference, context.length)
}
\arguments{
\item{vcf.file}{VCF file including path.}

\item{bsg}{BSgenome object.}

\item{reference}{A data.frame with the following columns: Mutation_Type, and names of SBS signatures (default: a data.frame returned from \code{\link{GetPcawgSbsSignaturesData}}).}

\item{context.length}{Number of context nucleotides.}
}
\value{
A list with the following items:
\item{df.sbs}{A data.frame of SBS data.}
\item{df.sbs.frequencies}{A data.frame of SBS mutation type frequencies.}
}
\description{
This function prepares a SBS VCF file and
returns relevant data to run IdentifySignatures function.
}
