% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf_parsing.R
\name{PrepareVcfFile}
\alias{PrepareVcfFile}
\title{Prepares a VCF file}
\usage{
PrepareVcfFile(vcf.file)
}
\arguments{
\item{vcf.file}{VCF file including path.}
}
\value{
A data.frame with the following items:
\item{Chr}{Chromosome name.}
\item{Pos}{Genomic position.}
\item{Ref}{Genomic position.}
\item{Alt}{Genomic position.}
}
\description{
This function prepares a VCF file.
}
