% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_dbs.R
\name{PreprocessDbs}
\alias{PreprocessDbs}
\title{Preprocess doublet base substitutions}
\usage{
PreprocessDbs(df, bsg, analyze.variants)
}
\arguments{
\item{df}{A data.frame with the following columns: Chr, Pos, Ref, Alt.}

\item{bsg}{BSgenome object.}

\item{analyze.variants}{A boolean value that indicates whether variant-level signature analysis should be performed.}
}
\value{
A data.frame with the following columns:
\item{Chr}{Chromosome name.}
\item{Pos}{Genomic position.}
\item{Ref}{Reference nucleotide sequence.}
\item{Alt}{Alternate nucleotide sequence.}
\item{Mutation_Type}{One of the 96 mutation sub types (e.g. AC>CA).}
}
\description{
This function preprocesses a data.frame of DBS.
}
