% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_indel.R
\name{PreprocessIndels}
\alias{PreprocessIndels}
\title{Preprocess indels}
\usage{
PreprocessIndels(df, bsg, padding.len, analyze.variants)
}
\arguments{
\item{df}{A data.frame with the following columns: Chr, Pos, Ref, Alt.}

\item{bsg}{BSgenome object.}

\item{padding.len}{Number of bases to fetch upstream and downstream of each indel.}

\item{analyze.variants}{A boolean value that indicates whether variant-level signature analysis should be performed.}
}
\value{
A data.frame with the following columns:
\item{Chr}{Chromosome name.}
\item{Pos}{Genomic position.}
\item{Ref}{Reference nucleotide sequence.}
\item{Alt}{Alternate nucleotide sequence.}
\item{Type}{Either 'INS' (insertion) or 'DEL' (deletion).}
\item{Len}{Length (number of bases) of this indel.}
\item{Mutated_Seq}{Nucleotide sequence of this indel.}
\item{Upstream_Seq}{Upstream nucleotide sequence of this indel.}
\item{Upstream_Start}{Upstream nucleotide sequence start position.}
\item{Upstream_End}{Upstream nucleotide sequence end position.}
\item{Downstream_Seq}{Downstream nucleotide sequence of this indel.}
\item{Downstream_Start}{Downstream nucleotide sequence start position..}
\item{Downstream_End}{Downstream nucleotide sequence end position.}
}
\description{
This function preprocesses a data.frame of indels.
}
