% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_sbs.R
\name{PreprocessSbs}
\alias{PreprocessSbs}
\title{Preprocess single base substitutions}
\usage{
PreprocessSbs(df, bsg, context.length, analyze.variants)
}
\arguments{
\item{df}{A data.frame with the following columns: Chr, Pos, Ref, Alt.}

\item{bsg}{BSgenome object.}

\item{context.length}{Number of context nucleotides.}

\item{analyze.variants}{A boolean value that indicates whether variant-level signature analysis should be performed.}
}
\value{
A data.frame with the following columns:
\item{Chr}{Chromosome name.}
\item{Pos}{Genomic position.}
\item{Ref}{Reference nucleotide sequence.}
\item{Alt}{Alternate nucleotide sequence.}
\item{Type}{One of =the six mutation types (e.g. C>A).}
\item{Sub_Type}{One of the 32 trinucleotide subtypes (e.g. ACA).}
\item{Somatic_Mutation_Type}{One of the 96 mutation sub types (e.g. A[C>A]A).}
\item{Upstream_Seq}{Upstream nucleotide sequence of this SBS.}
\item{Downstream_Seq}{Downstream nucleotide sequence of this SBS.}
}
\description{
This function preprocesses a data.frame of SBS.
}
