% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{PrintLog}
\alias{PrintLog}
\title{Prints log}
\usage{
PrintLog(message, type = "INFO")
}
\arguments{
\item{message}{String value message to print along with log type and date.}

\item{type}{String value that represents type of this message. 'INFO' by default.}
}
\description{
This function prints a log message.
}
