% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_group.R
\name{SortDataFrameForStackedBarPlot}
\alias{SortDataFrameForStackedBarPlot}
\title{Sorts a data.frame before plotting a stacked barplot}
\usage{
SortDataFrameForStackedBarPlot(df, ordered.features, x.axis.var)
}
\arguments{
\item{df}{A data.frame.}

\item{ordered.features}{An ordered character vector of columns to sort in df.}

\item{x.axis.var}{String value for x axis variable.}
}
\value{
A sorted data.frame.
}
\description{
This function sorts a data.frame before plotting a stacked barplot.
}
