% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{UnwrapIdentifiedModel}
\alias{UnwrapIdentifiedModel}
\title{Unwraps an identified model}
\usage{
UnwrapIdentifiedModel(identified.model)
}
\arguments{
\item{identified.model}{A data.frame returned from
\code{\link{IdentifySbsSignatures}}, \code{\link{IdentifyDbsSignatures}}, or \code{\link{IdentifyIndelSignatures}}.}
}
\value{
A data.frame.
}
\description{
This function returns an unwrapped list of an identified model.
}
