% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{WrapIdentifiedModel}
\alias{WrapIdentifiedModel}
\title{Wraps an identified model}
\usage{
WrapIdentifiedModel(
  mutations.count,
  signatures,
  signatures.weights,
  mutation.types,
  mutation.types.groups,
  observed.spectrum,
  reconstructed.spectrum,
  residual.spectrum,
  cosine.similarity,
  rss,
  rss.normalized,
  bic
)
}
\arguments{
\item{mutations.count}{Mutations count.}

\item{signatures}{A character vector of signature names.}

\item{signatures.weights}{A numeric vector of signature weights.}

\item{mutation.types}{A character vector of mutation types.}

\item{mutation.types.groups}{A character vector of mutation type groups.}

\item{observed.spectrum}{A numeric vector of observed spectrum.}

\item{reconstructed.spectrum}{A numeric vector of reconstructed spectrum.}

\item{residual.spectrum}{A numeric vector of residual spectrum.}

\item{cosine.similarity}{Cosine similiarity score.}

\item{rss}{Residual sum of squares.}

\item{rss.normalized}{Normalized residual sum of squares.}

\item{bic}{Bayesian information criterion.}
}
\value{
A data.frame with the following columns:
}
\description{
This function returns a (wrapped) data.frame of an identified model
}
